NCI-Agency/anet#18 Move actions and reducers tests to tests folder
